+++ src/panels/config/devices/device-detail/integration-elements/mqtt/mqtt-rx-messages.ts
@@ -1,0 +1,13 @@
+import { customElement } from "lit/decorators";
+import { MQTTMessages } from "./mqtt-messages";
+
+@customElement("mqtt-rx-messages")
+class MQTTRxMessages extends MQTTMessages {
+  direction = "Received";
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    "mqtt-rx-messages": MQTTRxMessages;
+  }
+}
