+++ src/panels/config/devices/device-detail/integration-elements/mqtt/mqtt-tx-messages.ts
@@ -1,0 +1,13 @@
+import { customElement } from "lit/decorators";
+import { MQTTMessages } from "./mqtt-messages";
+
+@customElement("mqtt-tx-messages")
+class MQTTTxMessages extends MQTTMessages {
+  direction = "Transmitted";
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    "mqtt-tx-messages": MQTTTxMessages;
+  }
+}
